requirements.txt
torch>=2.0.0
transformers>=4.38.0
peft>=0.9.0
accelerate>=0.27.0
chronos-forecasting>=1.0.0
statsmodels>=0.14.0
statsforecast>=1.7.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
pyyaml>=6.0
# Optional: second foundation models (install separately)
# pip install lag-llama
# pip install timesfm
